df1 egzas update , delete

+++ df1/app/Http/Controllers/CategoryController.php
@@ -93,6 +93,9 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->movies()->count()) {
+            return 'Negalima.';
+        }
         $category->delete();
         return redirect()->route('c_index');
     }
